move adblock detection to it's own hook
