+++ dotcom-rendering/src/lib/useAdBlockAsk.ts
@@ -3,72 +3,12 @@ import { getConsentFor, onConsentChange } from '@guardian/libs';
 import { useEffect, useState } from 'react';
 import { adFreeDataIsPresent } from '../client/userFeatures/user-features-lib';
 import { useAB } from './useAB';
+import { useDetectAdBlock } from './useDetectAdBlock';
 
 const useIsInAdBlockAskVariant = (): boolean => {
 	const abTestAPI = useAB()?.api;
 	const isInVariant = !!abTestAPI?.isUserInVariant('AdBlockAsk', 'variant');
 	return isInVariant;
-};
-
-/**
- * Make a HEAD request to a URL that is typically blocked by ad-blockers
- */
-const requestDoubleclick = async (timeoutMs: number) => {
-	try {
-		const response = await fetch('https://www3.doubleclick.net', {
-			method: 'HEAD',
-			mode: 'no-cors',
-			cache: 'no-store',
-			signal: AbortSignal.timeout(timeoutMs),
-		});
-
-		// A redirect is another clue we may be being ad-blocked
-		if (response.redirected) {
-			return false;
-		}
-
-		return true;
-	} catch (err) {
-		return false;
-	}
-};
-
-/**
- * Make a HEAD request to a URL that should succeed, even when using an
- * ad-blocker
- */
-const requestGuardian = async (timeoutMs: number) => {
-	try {
-		await fetch('https://www.theguardian.com', {
-			method: 'HEAD',
-			mode: 'no-cors',
-			cache: 'no-store',
-			signal: AbortSignal.timeout(timeoutMs),
-		});
-		return true;
-	} catch (err) {
-		return false;
-	}
-};
-
-/**
- * Attempt to detect presence of an ad-blocker
- *
- * This implementation of this is likely to be tweaked before launching the test
- * proper
- */
-const detectByRequests = async () => {
-	const [doubleclickSuccess, guardianSuccess] = await Promise.all([
-		requestDoubleclick(1000),
-		/**
-		 * We set this request with a much smaller timeout than the one we
-		 * expect to be ad-blocked. This should reduce the chance that request
-		 * fails and this one succeeds due to poor network connectivity
-		 */
-		requestGuardian(250),
-	]);
-
-	return !doubleclickSuccess && guardianSuccess;
 };
 
 export const useAdblockAsk = ({
@@ -81,7 +21,7 @@ export const useAdblockAsk = ({
 	isPaidContent: boolean;
 }): boolean => {
 	const isInVariant = useIsInAdBlockAskVariant();
-	const [adBlockerDetected, setAdBlockerDetected] = useState<boolean>(false);
+	const adBlockerDetected = useDetectAdBlock();
 	const [isAdFree, setIsAdFree] = useState<boolean>(false);
 	const [hasConsentForGoogletag, setHasConsentForGoogletag] = useState(false);
 
@@ -107,36 +47,22 @@ export const useAdblockAsk = ({
 	}, []);
 
 	useEffect(() => {
-		const makeRequest = async () => {
-			// Only perform the detection check in the variant of the AB test and if we haven't already detected an ad-blocker
-			if (isInVariant && !adBlockerDetected) {
-				EventTimer.get().setProperty('detectedAdBlocker', false);
+		// Only perform the detection check in the variant of the AB test, if we haven't already detected an ad-blocker and the reader/content is eligible for displaying such a message
+		if (isInVariant) {
+			EventTimer.get().setProperty('didDisplayAdBlockAsk', false);
 
-				if (await detectByRequests()) {
-					setAdBlockerDetected(true);
-
-					// Is the reader/content eligible for displaying such a message
-					if (canDisplayAdBlockAsk) {
-						// Some ad-blockers will remove slots from the DOM, while others don't
-						// This clean-up ensures that any space we've reserved for an ad is removed,
-						// in order to properly layout the ask.
-						document
-							.getElementById(slotId)
-							?.closest('.ad-slot-container')
-							?.remove();
-						EventTimer.get().setProperty(
-							'didDisplayAdBlockAsk',
-							true,
-						);
-					}
-
-					// Record ad block detection in commercial metrics
-					EventTimer.get().setProperty('detectedAdBlocker', true);
-				}
+			if (adBlockerDetected && canDisplayAdBlockAsk) {
+				// Some ad-blockers will remove slots from the DOM, while others don't
+				// This clean-up ensures that any space we've reserved for an ad is removed,
+				// in order to properly layout the ask.
+				document
+					.getElementById(slotId)
+					?.closest('.ad-slot-container')
+					?.remove();
+				EventTimer.get().setProperty('didDisplayAdBlockAsk', true);
 			}
-		};
-		void makeRequest();
+		}
 	}, [isInVariant, adBlockerDetected, slotId, canDisplayAdBlockAsk]);
 
-	return adBlockerDetected;
+	return adBlockerDetected && canDisplayAdBlockAsk && isInVariant;
 };
